# Optimization Attempt 1: ByteData