--- a/pubspec.yaml
+++ b/pubspec.yaml
@@ -13,3 +13,7 @@ dependencies:
 dev_dependencies:
   lints: ^2.0.0
   test: ^1.16.0
+
+dependency_overrides:
+  dart_z80: 
+    path: ../z80