--- a/test/spectrum_test.dart
+++ b/test/spectrum_test.dart
@@ -1,0 +1,15 @@
+import 'dart:io';
+
+import 'package:test/test.dart';
+import 'package:zxspectrum/zxspectrum.dart';
+
+void main() {
+  test('Reset is successful', () {
+    final rom = File('roms/48.rom').readAsBytesSync();
+    final spectrum = Spectrum(rom);
+    spectrum.memory.load(0x4000, List.generate(0x10000 - 0x4000, (index) => 0));
+    spectrum.reset();
+    final resetMemory = spectrum.memory.toList();
+    expect(resetMemory.sublist(0x4000), everyElement(0));
+  });
+}