Before:
```
Measuring the speed of 500 boot times.
Median boot time: 43.9ms.
Maximum boot time: 48.2ms.
Minimum boot time: 43.4ms.
```

After:
```
Measuring the speed of 500 boot times.
Median boot time: 44.5ms.
Maximum boot time: 48.8ms.
Minimum boot time: 43.4ms.
```